requirements.txt
Django>=4.2,<5
djangorestframework
djangorestframework-simplejwt
django-filter
psycopg2-binary
drf-yasg
python-dotenv
